+++ drivers/uniswap.js
@@ -8,27 +8,33 @@ const { parseToFloat } = require('../lib/utils.js');
  * @augments Driver
  */
 class Uniswap extends Driver {
+  constructor() {
+    super({
+      requires: {
+        key: true,
+      },
+    });
+  }
+
   /**
    * @augments Driver.fetchTickers
    * @returns {Promise.Array<Ticker>} Returns a promise of an array with tickers.
    */
   async fetchTickers() {
-    const pairs = await request('https://api.uniswap.info/v1/tickers');
+    const { results: tickers } = await request(`https://api.blocklytics.org/pools/v0/exchanges?platform=uniswap&key=${this.key}`);
 
-    return Object.keys(pairs).map((pair) => {
-      const ticker = pairs[pair];
-
-      return new Ticker({
-        base: ticker.base_symbol,
-        baseName: ticker.base_name,
-        quote: ticker.quote_symbol,
-        quoteName: ticker.quote_name,
-        quoteReference: ticker.quote_token_address,
-        close: parseToFloat(ticker.last_price),
-        baseVolume: parseToFloat(ticker.base_volume),
-        quoteVolume: parseToFloat(ticker.quote_volume),
-      });
-    });
+    return tickers.map((ticker) => new Ticker({
+      // Warning: quote and base are inverted
+      base: ticker.tokenSymbol,
+      baseName: ticker.tokenName,
+      baseReference: ticker.token,
+      quote: ticker.baseSymbol,
+      quoteName: ticker.baseName,
+      quoteReference: ticker.base,
+      close: parseToFloat(1 / ticker.basePrice),
+      baseVolume: parseToFloat(ticker.tokenVolume),
+      quoteVolume: parseToFloat(ticker.baseVolume),
+    }));
   }
 }